+++ packages/bpk-component-logo/index.js
@@ -1,7 +1,0 @@
-import BpkCloudLogo from './src/BpkCloudLogo';
-import BpkInlineLogo from './src/BpkInlineLogo';
-import BpkStackedLogo from './src/BpkStackedLogo';
-import BpkTianxunLogo from './src/BpkTianxunLogo';
-import BpkTianxunStackedLogo from './src/BpkTianxunStackedLogo';
-
-export { BpkCloudLogo, BpkInlineLogo, BpkStackedLogo, BpkTianxunLogo, BpkTianxunStackedLogo };