+++ packages/bpk-docs/src/pages/LogosPage/index.js
@@ -1,3 +1,0 @@
-import page from './LogosPage';
-
-export default page;